pyproject.toml
[project]
name = "watch-with-friends"
version = "0.1.0"
description = "Self-hosted web-app that let's you watch movides together using Yandex.Disk"
requires-python = ">=3.12"
dependencies = [
    "nicegui>=2.21.1",
    "aiohttp>=3.12.14",
    "aiofiles>=24.1.0",
    "yndx-disk>=0.3.2",
    "cachetools>=6.1.0",
    "cachetools-async>=0.0.5",
    "orjson>=3.11.0",
    "bcrypt>=4.3.0",
    "pyjwt>=2.10.1",
]
